+++ src/pyramid_bimt/views/__init__.py
@@ -47,6 +47,7 @@ class FormView(BaseFormView):
         result = super(FormView, self).__call__()
         if isinstance(result, dict):  # pragma: no branch
             result['title'] = self.title
+            self.request.layout_manager.layout.title = self.title
             result['description'] = getattr(self, 'description', None)
         return result