+++ src/pyramid_bimt/views/group.py
@@ -33,6 +33,7 @@ class GroupView(object):
     )
     def list(self):
         self.request.layout_manager.layout.hide_sidebar = True
+        self.request.layout_manager.layout.title = u'Groups'
         return {
             'groups': Group.get_all().all(),
         }